+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @demex-sdk/core
 
+## 0.0.6
+
+### Patch Changes
+
+- Added EIP-712 signing functionality
+- Updated dependencies
+  - @demex-sdk/codecs@0.0.6
+
 ## 0.0.3
 
 ### Patch Changes